backend/requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
google-auth==2.23.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
